Fix: Implementation of a compoenent for how it work section and adding into the landing page
